@@ -71,33 +74,38 @@
     "@types/redux-logger": "^3.0.9",
     "@typescript-eslint/eslint-plugin": "^4.13.0",
     "@typescript-eslint/parser": "^4.13.0",
+    "@vitejs/plugin-basic-ssl": "^1.0.2",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "css-mediaquery": "^0.1.2",
     "customize-cra": "^1.0.0",
     "dcl-tslint-config-standard": "^3.0.0",
+    "dotenv": "^16.3.1",
     "http-proxy-middleware": "^2.0.6",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "node-fetch": "^2.7.0",
     "prettier": "^1.19.1",
-    "react-app-rewired": "^2.2.1",
-    "react-error-overlay": "^6.0.9",
     "redux-saga-test-plan": "^4.0.3",
+    "rollup-plugin-polyfill-node": "^0.13.0",
+    "ts-node": "^10.9.2",
     "tslint": "^5.20.1",
     "tslint-plugin-prettier": "^2.0.1",
     "tslint-react": "^4.1.0",
     "typechain": "^8.0.0",
-    "typescript": "^4.9.3"
+    "typescript": "^5.3.3",
+    "vite": "^5.0.11"
   },
   "scripts": {
-    "start": "react-app-rewired --openssl-legacy-provider start",
-    "prebuild": "node scripts/prebuild.js",
-    "build": "CI=false react-scripts --openssl-legacy-provider build",
-    "test": "react-app-rewired test",
+    "start": "vite",
+    "preview": "vite preview",
+    "prebuild": "node scripts/prebuild.cjs",
+    "build": "tsc && node --max_old_space_size=8192 ./node_modules/vite/bin/vite.js build",
+    "test": "jest",
     "test:coverage": "npm run test -- --coverage",
-    "eject": "react-app-rewired eject",
     "lint": "tslint -p tsconfig.json",
     "lint:fix": "tslint -p tsconfig.json --fix",
     "generate-contracts": "typechain --target ethers-v5 --out-dir src/contracts src/contracts/*.json"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "prettier": {
     "printWidth": 80,
