@@ -1,3 +1,4 @@
+/* eslint-disable */
 const fs = require('fs')
 const dotenv = require('dotenv')
 
@@ -10,24 +11,20 @@ if (fs.existsSync('.env')) {
   Object.assign(ENV_CONTENT, dotenv.parse(fs.readFileSync('.env')))
 }
 const packageJson = JSON.parse(fs.readFileSync('./package.json').toString())
-const publicPackageJson = JSON.parse(
-  fs.readFileSync('./public/package.json').toString()
-)
+const publicPackageJson = JSON.parse(fs.readFileSync('./public/package.json').toString())
 
 // set version
-ENV_CONTENT['REACT_APP_WEBSITE_VERSION'] = packageJson.version
+ENV_CONTENT['VITE_REACT_APP_WEBSITE_VERSION'] = packageJson.version
 publicPackageJson.version = packageJson.version
 
 // set public url
 Object.assign(ENV_CONTENT, getPublicUrls())
-packageJson.homepage = ENV_CONTENT['PUBLIC_URL']
+packageJson.homepage = ENV_CONTENT['VITE_BASE_URL']
 publicPackageJson.homepage = packageJson.homepage
 if (packageJson.homepage) {
   // github action outputs. Do not touch.
   console.log('::set-output name=public_url::' + packageJson.homepage)
-  console.log(
-    '::set-output name=public_path::' + new URL(packageJson.homepage).pathname
-  )
+  console.log('::set-output name=public_path::' + new URL(packageJson.homepage).pathname)
 }
 
 // log stuff
@@ -41,10 +38,7 @@ fs.writeFileSync(
     .join('\n') + '\n'
 )
 fs.writeFileSync('./package.json', JSON.stringify(packageJson, null, 2))
-fs.writeFileSync(
-  './public/package.json',
-  JSON.stringify(publicPackageJson, null, 2)
-)
+fs.writeFileSync('./public/package.json', JSON.stringify(publicPackageJson, null, 2))
 
 // public url logic
 function getPublicUrls() {
@@ -61,12 +55,12 @@ function getPublicUrls() {
     const cdnUrl = `https://cdn.decentraland.org/${publicPackageJson.name}/${publicPackageJson.version}`
     console.log(`Using CDN as public url: "${cdnUrl}"`)
     return {
-      PUBLIC_URL: cdnUrl
+      VITE_BASE_URL: cdnUrl
     }
   }
   // localhost
-  console.log(`Using empty pubic url`)
+  console.log('Using empty pubic url')
   return {
-    PUBLIC_URL: ``
+    VITE_BASE_URL: ''
   }
 }
