@@ -1,10 +1,10 @@
 import { connect } from 'react-redux'
+import { openModal } from 'decentraland-dapps/dist/modules/modal/actions'
 import { RootState } from '../../../modules/reducer'
 import { getWallet } from '../../../modules/wallet/selectors'
 import { getCurrentOrder } from '../../../modules/order/selectors'
 import { getNFTBids } from '../../../modules/ui/nft/bid/selectors'
 import { NFT } from '../../../modules/nft/types'
-import { openModal } from '../../../modules/modal/actions'
 import Actions from './Actions'
 import { MapDispatch, MapDispatchProps, MapStateProps } from './Actions.types'
 
