@@ -3,6 +3,7 @@ import { replace } from 'connected-react-router'
 import { Order } from '@dcl/schemas'
 import { openModal } from 'decentraland-dapps/dist/modules/modal/actions'
 import { getOpenModals } from 'decentraland-dapps/dist/modules/modal/selectors'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
 import {
   MapDispatchProps,
   MapDispatch,
@@ -13,7 +14,7 @@ import { buyItemWithCardRequest } from '../../../../modules/item/actions'
 import { Asset } from '../../../../modules/asset/types'
 import { RootState } from '../../../../modules/reducer'
 import { getIsBuyCrossChainEnabled } from '../../../../modules/features/selectors'
-import { getWallet, isConnecting } from '../../../../modules/wallet/selectors'
+import { getWallet } from '../../../../modules/wallet/selectors'
 import BuyNFTButtons from './BuyNFTButtons'
 
 const mapState = (state: RootState): MapStateProps => ({
