@@ -2,9 +2,9 @@ import { connect } from 'react-redux'
 import { Network } from '@dcl/schemas'
 import { isLoadingType } from 'decentraland-dapps/dist/modules/loading/selectors'
 import { FETCH_APPLICATION_FEATURES_REQUEST } from 'decentraland-dapps/dist/modules/features/actions'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
 import { isLoadingFeatureFlags as getIsLoadingFeatureFlags } from '../../modules/features/selectors'
 import { RootState } from '../../modules/reducer'
-import { isConnecting } from '../../modules/wallet/selectors'
 import {
   fetchNFTRequest,
   FETCH_NFT_REQUEST,
