@@ -1,8 +1,9 @@
 import { Dispatch } from 'redux'
 import { connect } from 'react-redux'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
 import { goBack } from '../../modules/routing/actions'
 import { RootState } from '../../modules/reducer'
-import { getWallet, isConnecting } from '../../modules/wallet/selectors'
+import { getWallet } from '../../modules/wallet/selectors'
 import { MapDispatchProps, MapStateProps } from './ManageAssetPage.types'
 import { ManageAssetPage } from './ManageAssetPage'
 
