@@ -2,11 +2,9 @@ import { connect } from 'react-redux'
 import { push, replace } from 'connected-react-router'
 import { Network } from '@dcl/schemas'
 import { openModal } from 'decentraland-dapps/dist/modules/modal/actions'
-import {
-  getMana,
-  getWallet,
-  isConnecting
-} from '../../../modules/wallet/selectors'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
+
+import { getMana, getWallet } from '../../../modules/wallet/selectors'
 import { locations } from '../../../modules/routing/locations'
 import { Section } from '../../../modules/vendor/decentraland'
 import { BrowseOptions } from '../../../modules/routing/types'
