@@ -1,5 +1,6 @@
 import { connect } from 'react-redux'
-import { getWallet, isConnecting } from '../../modules/wallet/selectors'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
+import { getWallet } from '../../modules/wallet/selectors'
 import { RootState } from '../../modules/reducer'
 import ProtectedRoute from './ProtectedRoute'
 import { MapStateProps } from './ProtectedRoute.types'
