@@ -14,7 +14,8 @@ import {
   isContractAccountError
 } from '../../modules/transaction/utils'
 import { isLoadingType } from 'decentraland-dapps/dist/modules/loading/selectors'
-import { getWallet, isConnecting } from '../../modules/wallet/selectors'
+import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
+import { getWallet } from '../../modules/wallet/selectors'
 import {
   getContract,
   getHasFetched,
