@@ -27,7 +27,13 @@ function renderSuccessPage(
   preloadedNFTData?: Record<string, NFT>
 ): RenderResult {
   return renderWithProviders(
-    <SuccessPage isLoading={false} mintedTokenId={null} {...props} />,
+    <SuccessPage
+      profile={undefined}
+      onSetNameAsAlias={() => undefined}
+      isLoading={false}
+      mintedTokenId={null}
+      {...props}
+    />,
     {
       preloadedState: {
         nft: {
