@@ -12,6 +12,10 @@ import {
   NFTPurchase,
   PurchaseStatus
 } from 'decentraland-dapps/dist/modules/gateway/types'
+import {
+  closeModal,
+  openModal
+} from 'decentraland-dapps/dist/modules/modal/actions'
 import { NetworkGatewayType } from 'decentraland-ui'
 import { locations } from '../routing/locations'
 import { fetchSmartWearableRequiredPermissionsRequest } from '../asset/actions'
@@ -19,7 +23,6 @@ import { getWallet } from '../wallet/selectors'
 import { View } from '../ui/types'
 import { ItemAPI } from '../vendor/decentraland/item/api'
 import { CatalogAPI } from '../vendor/decentraland/catalog/api'
-import { closeModal, openModal } from '../modal/actions'
 import {
   buyAssetWithCard,
   BUY_NFTS_WITH_CARD_EXPLANATION_POPUP_KEY
