@@ -1,5 +1,10 @@
 import { RentalListing } from '@dcl/schemas'
 import { getOpenModals } from 'decentraland-dapps/dist/modules/modal/selectors'
+import {
+  closeAllModals,
+  closeModal,
+  openModal
+} from 'decentraland-dapps/dist/modules/modal/actions'
 import { expectSaga } from 'redux-saga-test-plan'
 import { select } from 'redux-saga/effects'
 import { NFT } from '../nft/types'
@@ -20,7 +25,6 @@ import {
   UPDATE_LIST_SUCCESS
 } from '../favorites/actions'
 import { UpsertRentalOptType } from '../rental/types'
-import { closeAllModals, closeModal, openModal } from './actions'
 import { modalSaga } from './sagas'
 
 describe.each([
