@@ -67,6 +67,7 @@ let rentalListing: RentalListing
 const purchase: NFTPurchase = {
   address: 'anAddress',
   id: 'anId',
+  paymentMethod: 'aPaymentMethod',
   network: Network.ETHEREUM,
   timestamp: 1671028355396,
   status: PurchaseStatus.PENDING,
@@ -84,6 +85,7 @@ const purchase: NFTPurchase = {
 beforeEach(() => {
   state = {
     data: {},
+    dataLegacy: {},
     loading: [],
     error: null
   }
@@ -119,7 +121,7 @@ requestActions.forEach(action => {
 
 const successActions = [
   createOrderSuccess(nft, Number(order.price), order.expiresAt, txHash),
-  executeOrderSuccess(),
+  executeOrderSuccess(txHash, nft),
   executeOrderWithCardSuccess(purchase, nft, txHash),
   cancelOrderSuccess(order, nft, txHash)
 ]
