@@ -11,6 +11,7 @@ import { AuthIdentity } from 'decentraland-crypto-fetch'
 import { getConnectedProvider } from 'decentraland-dapps/dist/lib/eth'
 import { waitForTx } from 'decentraland-dapps/dist/modules/transaction/utils'
 import { sendTransaction } from 'decentraland-dapps/dist/modules/wallet/utils'
+import { closeModal } from 'decentraland-dapps/dist/modules/modal/actions'
 import {
   ContractData,
   ContractName,
@@ -21,7 +22,6 @@ import { expectSaga } from 'redux-saga-test-plan'
 import { throwError } from 'redux-saga-test-plan/providers'
 import { delay, take } from 'redux-saga/effects'
 import { getCurrentIdentity } from '../identity/selectors'
-import { closeModal } from '../modal/actions'
 import { FETCH_NFT_SUCCESS } from '../nft/actions'
 import { getCurrentNFT } from '../nft/selectors'
 import { NFT } from '../nft/types'
